+++ src/PhaseDrivenSequencer.cpp
@@ -25,6 +25,7 @@ struct PhaseDrivenSequencer : HCVModule
         PHASOR_INPUT,
         STEPSCV_INPUT,
         WIDTHCV_INPUT,
+        RUN_INPUT,
 		NUM_INPUTS
 	};
 	enum OutputIds
@@ -44,6 +45,7 @@ struct PhaseDrivenSequencer : HCVModule
         SLEW_LIGHT,
         SH_LIGHT,
         GATES_LIGHT,
+        RUN_LIGHT,
         NUM_LIGHTS
 	};
 
@@ -173,7 +175,13 @@ void PhaseDrivenSequencer::process(const ProcessArgs &args)
         float pulseWidth = widthKnob + (widthDepth * inputs[WIDTHCV_INPUT].getPolyVoltage(i));
         pulseWidth = clamp(pulseWidth, -5.0f, 5.0f) * 0.1f + 0.5f;
 
-        const float phasorIn = inputs[PHASOR_INPUT].getPolyVoltage(i);
+        bool active = true;
+        if(inputs[RUN_INPUT].isConnected())
+        {
+            active = inputs[RUN_INPUT].getPolyVoltage(i) >= 1.0f;
+        }
+
+        const float phasorIn = active ? inputs[PHASOR_INPUT].getPolyVoltage(i) : 0.0f;
         float normalizedPhasor = scaleAndWrapPhasor(phasorIn);
 
         stepDetectors[i].setNumberSteps(numSteps);
@@ -211,7 +219,7 @@ void PhaseDrivenSequencer::process(const ProcessArgs &args)
         else
         {
             const float gate = fractionalIndex < pulseWidth ? HCV_PHZ_GATESCALE : 0.0f;
-            outputs[GATES_OUTPUT].setVoltage(gates[currentIndex] ? gate : 0.0f, i);
+            outputs[GATES_OUTPUT].setVoltage(gates[currentIndex] && active ? gate : 0.0f, i);
 
             bool trigger = gate && gates[currentIndex];
             outputs[TRIGS_OUTPUT].setVoltage(triggers[i].process(trigger) ? HCV_PHZ_GATESCALE : 0.0f, i);
@@ -230,7 +238,12 @@ void PhaseDrivenSequencer::process(const ProcessArgs &args)
         if(i == 0) lightIndex = currentIndex;
     }
 
-    bool isPlaying = slopeDetectors[0].isPhasorAdvancing();
+    bool active = true;
+    if(inputs[RUN_INPUT].isConnected())
+    {
+        active = inputs[RUN_INPUT].getPolyVoltage(0) >= 1.0f;
+    }
+    bool isPlaying = slopeDetectors[0].isPhasorAdvancing() && active;
 
     // Gate lights
     for (int i = 0; i < NUM_STEPS; i++) 
@@ -240,6 +253,7 @@ void PhaseDrivenSequencer::process(const ProcessArgs &args)
         lights[GATE_LIGHTS + 3 * i + 2].setSmoothBrightness(isPlaying && lightIndex == i, args.sampleTime); //blue
     }
 
+    lights[RUN_LIGHT].setBrightness(active ? 1.0f : 0.0f);
     setLightFromOutput(STEPS_LIGHT, STEPS_OUTPUT, 0.2f);
     setLightFromOutput(SLEW_LIGHT, SLEW_OUTPUT, 0.2f);
     setLightFromOutput(SH_LIGHT, SH_OUTPUT, 0.2f);
@@ -259,20 +273,24 @@ PhaseDrivenSequencerWidget::PhaseDrivenSequencerWidget(PhaseDrivenSequencer *mod
     createParamComboVertical(15, knobY, PhaseDrivenSequencer::WIDTH_PARAM, PhaseDrivenSequencer::WIDTHCV_PARAM, PhaseDrivenSequencer::WIDTHCV_INPUT);
     createParamComboVertical(70, knobY, PhaseDrivenSequencer::STEPS_PARAM, PhaseDrivenSequencer::STEPSCV_PARAM, PhaseDrivenSequencer::STEPSCV_INPUT);
     
-    createHCVSwitchVert(89, 255, PhaseDrivenSequencer::DETECTION_PARAM);
+    createHCVSwitchVert(53, 208, PhaseDrivenSequencer::DETECTION_PARAM);
 
     //////INPUTS//////
     int jackX = 20;
-    createInputPort(jackX, 248, PhaseDrivenSequencer::PHASOR_INPUT);
+    int jackX2 = 78;
+    int inputY = 248;
+    createInputPort(jackX, inputY, PhaseDrivenSequencer::PHASOR_INPUT);
+    createInputPort(jackX2, inputY, PhaseDrivenSequencer::RUN_INPUT);
 
     //////OUTPUTS/////
     int spacing = 45;
     int outJackX = 18;
-    createOutputPort(outJackX, 310, PhaseDrivenSequencer::STEPS_OUTPUT);
-    createOutputPort(outJackX + spacing, 310, PhaseDrivenSequencer::SLEW_OUTPUT);
-    createOutputPort(outJackX + spacing*2, 310, PhaseDrivenSequencer::SH_OUTPUT);
-    createOutputPort(outJackX + spacing*3, 310, PhaseDrivenSequencer::GATES_OUTPUT);
-    createOutputPort(outJackX + spacing*4, 310, PhaseDrivenSequencer::TRIGS_OUTPUT);
+    int outY = 310;
+    createOutputPort(outJackX, outY, PhaseDrivenSequencer::STEPS_OUTPUT);
+    createOutputPort(outJackX + spacing, outY, PhaseDrivenSequencer::SLEW_OUTPUT);
+    createOutputPort(outJackX + spacing*2, outY, PhaseDrivenSequencer::SH_OUTPUT);
+    createOutputPort(outJackX + spacing*3, outY, PhaseDrivenSequencer::GATES_OUTPUT);
+    createOutputPort(outJackX + spacing*4, outY, PhaseDrivenSequencer::TRIGS_OUTPUT);
 
     for (int i = 0; i < PhaseDrivenSequencer::NUM_STEPS; i++)
     {
@@ -287,11 +305,13 @@ PhaseDrivenSequencerWidget::PhaseDrivenSequencerWidget(PhaseDrivenSequencer *mod
         addParam(createLightParamCentered<VCVLightBezel<RedGreenBlueLight>>(Vec(buttonX, buttonY), 
         module, PhaseDrivenSequencer::GATE_PARAMS + i, PhaseDrivenSequencer::GATE_LIGHTS + 3 * i));
     }
+
+    createHCVGreenLightForJack(jackX2, inputY, PhaseDrivenSequencer::RUN_LIGHT);
     
-    createHCVRedLightForJack(outJackX, 310, PhaseDrivenSequencer::STEPS_LIGHT);
-    createHCVRedLightForJack(outJackX + spacing, 310, PhaseDrivenSequencer::SLEW_LIGHT);
-    createHCVRedLightForJack(outJackX + spacing*2, 310, PhaseDrivenSequencer::SH_LIGHT);
-    createHCVRedLightForJack(outJackX + spacing*3, 310, PhaseDrivenSequencer::GATES_LIGHT);
+    createHCVRedLightForJack(outJackX, outY, PhaseDrivenSequencer::STEPS_LIGHT);
+    createHCVRedLightForJack(outJackX + spacing, outY, PhaseDrivenSequencer::SLEW_LIGHT);
+    createHCVRedLightForJack(outJackX + spacing*2, outY, PhaseDrivenSequencer::SH_LIGHT);
+    createHCVRedLightForJack(outJackX + spacing*3, outY, PhaseDrivenSequencer::GATES_LIGHT);
 }
 
 Model *modelPhaseDrivenSequencer = createModel<PhaseDrivenSequencer, PhaseDrivenSequencerWidget>("PhaseDrivenSequencer");