+++ src/PhasorGates64.cpp
@@ -24,12 +24,15 @@ struct PhasorGates64 : HCVModule
         PHASOR_INPUT,
         STEPSCV_INPUT,
         WIDTHCV_INPUT,
+        RUN_INPUT,
 		NUM_INPUTS
 	};
 	enum OutputIds
 	{
         GATES_OUTPUT,
         TRIGS_OUTPUT,
+        PHASOR_OUTPUT,
+        GATES_NOT_OUTPUT,
 		NUM_OUTPUTS
     };
 
@@ -37,6 +40,9 @@ struct PhasorGates64 : HCVModule
     {
         ENUMS(GATE_LIGHTS, NUM_STEPS*3),
         GATE_OUT_LIGHT,
+        RUN_LIGHT,
+        PHASOR_LIGHT,
+        GATES_NOT_OUT_LIGHT,
         NUM_LIGHTS
 	};
 
@@ -81,19 +87,29 @@ struct PhasorGates64 : HCVModule
 
     void onReset() override
     {
+        resetGates();
+    }
+
+    void onRandomize() override 
+    {
+		randomizeGates();
+	}
+
+    void resetGates()
+    {
         for (int i = 0; i < NUM_STEPS; i++) 
         {
 			gates[i] = false;
 		}
     }
 
-    void onRandomize() override 
+    void randomizeGates()
     {
-		for (int i = 0; i < NUM_STEPS; i++) 
+        for (int i = 0; i < NUM_STEPS; i++) 
         {
 			gates[i] = random::get<bool>();
 		}
-	}
+    }
 
     json_t *dataToJson() override
     {
@@ -151,7 +167,13 @@ void PhasorGates64::process(const ProcessArgs &args)
         float pulseWidth = widthKnob + (widthDepth * inputs[WIDTHCV_INPUT].getPolyVoltage(i));
         pulseWidth = clamp(pulseWidth, -5.0f, 5.0f) * 0.1f + 0.5f;
 
-        const float phasorIn = inputs[PHASOR_INPUT].getPolyVoltage(i);
+        bool active = true;
+        if(inputs[RUN_INPUT].isConnected())
+        {
+            active = inputs[RUN_INPUT].getPolyVoltage(i) >= 1.0f;
+        }
+
+        const float phasorIn = active ? inputs[PHASOR_INPUT].getPolyVoltage(i) : 0.0f;
         float normalizedPhasor = scaleAndWrapPhasor(phasorIn);
 
         stepDetectors[i].setNumberSteps(numSteps);
@@ -166,36 +188,44 @@ void PhasorGates64::process(const ProcessArgs &args)
 
             if(slopeDetectors[i].isPhasorAdvancing() || !isZero)
             {
-                float gate;
-                if (reversePhasor) gate = (1.0f - fractionalIndex) < pulseWidth ? HCV_PHZ_GATESCALE : 0.0f;
-                else gate = fractionalIndex < pulseWidth ? HCV_PHZ_GATESCALE : 0.0f;
+                float fractionalOutput = reversePhasor ? (1.0f - fractionalIndex) : fractionalIndex;
+                float gate = fractionalOutput < pulseWidth ? HCV_PHZ_GATESCALE : 0.0f;
 
                 outputs[GATES_OUTPUT].setVoltage(gates[currentIndex] ? gate : 0.0f, i);
+                outputs[GATES_NOT_OUTPUT].setVoltage(!gates[currentIndex] ? gate : 0.0f, i);
 
                 bool trigger = gate && gates[currentIndex];
                 outputs[TRIGS_OUTPUT].setVoltage(triggers[i].process(trigger) ? HCV_PHZ_GATESCALE : 0.0f, i);
+                outputs[PHASOR_OUTPUT].setVoltage(gates[currentIndex] ? fractionalOutput * HCV_PHZ_UPSCALE : 0.0f, i);
             }
             else
             {
                 outputs[GATES_OUTPUT].setVoltage(0.0f, i);
+                outputs[GATES_NOT_OUTPUT].setVoltage(0.0f, i);
                 outputs[TRIGS_OUTPUT].setVoltage(0.0f, i);
+                outputs[PHASOR_OUTPUT].setVoltage(0.0f, i);
             }
         }
         else
         {
             const float gate = fractionalIndex < pulseWidth ? HCV_PHZ_GATESCALE : 0.0f;
-            outputs[GATES_OUTPUT].setVoltage(gates[currentIndex] ? gate : 0.0f, i);
+            outputs[GATES_OUTPUT].setVoltage(gates[currentIndex] && active ? gate : 0.0f, i);
+            outputs[GATES_NOT_OUTPUT].setVoltage(!gates[currentIndex] && active ? gate : 0.0f, i);
 
             bool trigger = gate && gates[currentIndex];
             outputs[TRIGS_OUTPUT].setVoltage(triggers[i].process(trigger) ? HCV_PHZ_GATESCALE : 0.0f, i);
+            outputs[PHASOR_OUTPUT].setVoltage(gates[currentIndex] ? fractionalIndex * HCV_PHZ_UPSCALE : 0.0f, i);
         }
-
-        
 
         if(i == 0) lightIndex = currentIndex;
     }
 
-    bool isPlaying = slopeDetectors[0].isPhasorAdvancing();
+    bool active = true;
+    if(inputs[RUN_INPUT].isConnected())
+    {
+        active = inputs[RUN_INPUT].getPolyVoltage(0) >= 1.0f;
+    }
+    bool isPlaying = slopeDetectors[0].isPhasorAdvancing() && active;
 
     // Gate buttons
     for (int i = 0; i < NUM_STEPS; i++) 
@@ -208,7 +238,10 @@ void PhasorGates64::process(const ProcessArgs &args)
         lights[GATE_LIGHTS + 3 * i + 2].setSmoothBrightness(isPlaying && lightIndex == i, args.sampleTime); //blue
     }
 
+    lights[RUN_LIGHT].setBrightness(active ? 1.0f : 0.0f);
     setLightFromOutput(GATE_OUT_LIGHT, GATES_OUTPUT);
+    setLightFromOutput(PHASOR_LIGHT, PHASOR_OUTPUT);
+    setLightFromOutput(GATES_NOT_OUT_LIGHT, GATES_NOT_OUTPUT);
 }
 
 struct PhasorGates64Widget : HCVModuleWidget { PhasorGates64Widget(PhasorGates64 *module); };
@@ -223,25 +256,37 @@ PhasorGates64Widget::PhasorGates64Widget(PhasorGates64 *module)
     createParamComboVertical(15, knobY, PhasorGates64::WIDTH_PARAM, PhasorGates64::WIDTHCV_PARAM, PhasorGates64::WIDTHCV_INPUT);
     createParamComboVertical(70, knobY, PhasorGates64::STEPS_PARAM, PhasorGates64::STEPSCV_PARAM, PhasorGates64::STEPSCV_INPUT);
     
-    createHCVSwitchVert(89, 255, PhasorGates64::DETECTION_PARAM);
+    createHCVSwitchVert(53, 208, PhasorGates64::DETECTION_PARAM);
 
     //////INPUTS//////
     int jackX = 20;
-    createInputPort(jackX, 248, PhasorGates64::PHASOR_INPUT);
+    int jackX2 = 78;
+    int jackX3 = 136;
+    int jackX4 = 194;
+    int inputY = 248;
+    createInputPort(jackX, inputY, PhasorGates64::PHASOR_INPUT);
+    createInputPort(jackX2, inputY, PhasorGates64::RUN_INPUT);
 
     //////OUTPUTS/////
-    createOutputPort(jackX, 310, PhasorGates64::GATES_OUTPUT);
-    createOutputPort(78, 310, PhasorGates64::TRIGS_OUTPUT);
+    int outputY = 310;
+    createOutputPort(jackX2, outputY, PhasorGates64::GATES_OUTPUT);
+    createOutputPort(jackX3, outputY, PhasorGates64::TRIGS_OUTPUT);
+    createOutputPort(jackX4, outputY, PhasorGates64::PHASOR_OUTPUT);
+    createOutputPort(jackX4 + 58.0f, outputY, PhasorGates64::GATES_NOT_OUTPUT);
 
     for (int i = 0; i < PhasorGates64::NUM_STEPS; i++)
     {
-        int buttonY = 55 + (i/8)*40;
+        int buttonY = 55 + (i/8)*30;
         int buttonX = 130 + 30 * (i%8);
         addParam(createLightParamCentered<VCVLightBezel<RedGreenBlueLight>>(Vec(buttonX, buttonY), 
         module, PhasorGates64::GATE_PARAMS + i, PhasorGates64::GATE_LIGHTS + 3 * i));
     }
     
-    createHCVRedLightForJack(jackX, 310, PhasorGates64::GATE_OUT_LIGHT);
+    createHCVGreenLightForJack(jackX2, inputY, PhasorGates64::RUN_LIGHT);
+
+    createHCVRedLightForJack(jackX2, outputY, PhasorGates64::GATE_OUT_LIGHT);
+    createHCVRedLightForJack(jackX4, outputY, PhasorGates64::PHASOR_LIGHT);
+    createHCVRedLightForJack(jackX4 + 58.0f, outputY, PhasorGates64::GATES_NOT_OUT_LIGHT);
 }
 
 Model *modelPhasorGates64 = createModel<PhasorGates64, PhasorGates64Widget>("PhasorGates64");